Cut out codes in the `helpers` block into a module
