+++ lib/resque/server.rb
@@ -1,6 +1,7 @@
 require 'sinatra/base'
 require 'tilt/erb'
 require 'resque'
+require 'resque/server_helper'
 require 'resque/version'
 require 'time'
 require 'yaml'
@@ -11,8 +12,6 @@ end
 
 module Resque
   class Server < Sinatra::Base
-    require 'resque/server/helpers'
-
     dir = File.dirname(File.expand_path(__FILE__))
 
     set :views,  "#{dir}/server/views"
@@ -26,121 +25,7 @@ module Resque
     set :static, true
 
     helpers do
-      include Rack::Utils
-      alias_method :h, :escape_html
-
-      def current_section
-        url_path request.path_info.sub('/','').split('/')[0].downcase
-      end
-
-      def current_page
-        url_path request.path_info.sub('/','')
-      end
-
-      def url_path(*path_parts)
-        [ url_prefix, path_prefix, path_parts ].join("/").squeeze('/')
-      end
-      alias_method :u, :url_path
-
-      def path_prefix
-        request.env['SCRIPT_NAME']
-      end
-
-      def class_if_current(path = '')
-        'class="current"' if current_page[0, path.size] == path
-      end
-
-      def tab(name)
-        dname = name.to_s.downcase
-        path = url_path(dname)
-        "<li #{class_if_current(path)}><a href='#{path.gsub(" ", "_")}'>#{name}</a></li>"
-      end
-
-      def tabs
-        Resque::Server.tabs
-      end
-
-      def url_prefix
-        Resque::Server.url_prefix
-      end
-
-      def redis_get_size(key)
-        case Resque.redis.type(key)
-        when 'none'
-          0
-        when 'hash'
-          Resque.redis.hlen(key)
-        when 'list'
-          Resque.redis.llen(key)
-        when 'set'
-          Resque.redis.scard(key)
-        when 'string'
-          Resque.redis.get(key).length
-        when 'zset'
-          Resque.redis.zcard(key)
-        end
-      end
-
-      def redis_get_value_as_array(key, start=0)
-        case Resque.redis.type(key)
-        when 'none'
-          []
-        when 'hash'
-          Resque.redis.hgetall(key).to_a[start..(start + 20)]
-        when 'list'
-          Resque.redis.lrange(key, start, start + 20)
-        when 'set'
-          Resque.redis.smembers(key)[start..(start + 20)]
-        when 'string'
-          [Resque.redis.get(key)]
-        when 'zset'
-          Resque.redis.zrange(key, start, start + 20)
-        end
-      end
-
-      def show_args(args)
-        Array(args).map do |a|
-          a.to_yaml
-        end.join("\n")
-      rescue
-        args.to_s
-      end
-
-      def worker_hosts
-        @worker_hosts ||= worker_hosts!
-      end
-
-      def worker_hosts!
-        hosts = Hash.new { [] }
-
-        Resque.workers.each do |worker|
-          host, _ = worker.to_s.split(':')
-          hosts[host] += [worker.to_s]
-        end
-
-        hosts
-      end
-
-      def partial?
-        @partial
-      end
-
-      def partial(template, local_vars = {})
-        @partial = true
-        erb(template.to_sym, {:layout => false}, local_vars)
-      ensure
-        @partial = false
-      end
-
-      def poll
-        if defined?(@polling) && @polling
-          text = "Last Updated: #{Time.now.strftime("%H:%M:%S")}"
-        else
-          text = "<a href='#{u(request.path_info)}.poll' rel='poll'>Live Poll</a>"
-        end
-        "<p class='poll'>#{text}</p>"
-      end
-
+      include Resque::ServerHelper
     end
 
     def show(page, layout = true)
